Move insteon base entity to separate module
