+++ homeassistant/components/insteon/entity.py
